Fix coverage crash when numeric columns are stored as TEXT
Editing a table in the app (st.data_editor) could persist a numeric column
as TEXT in SQLite, giving object dtype on read. This propagated through the
demand/required pipeline and crashed build_coverage at np.divide with
'Cannot cast ufunc divide output from dtype(O) to float64'.

Fix at the root in db.py: coerce_types() forces numeric dtype on every
measure / numeric dimension of a registered table, applied on both read_table
(repairs already-corrupted DBs in memory) and write_table (prevents future
TEXT storage). Plus a defensive float cast at the divide site in build_coverage.

Regression test corrupts 4 numeric columns to TEXT and asserts read coercion
and a successful coverage computation.

+++ staffing-engine/staffing/optimizer.py
@@ -190,10 +190,10 @@ def build_coverage(month: str, db_path=db.DB_PATH) -> dict:
         m[c] = m[c].fillna(0.0)
     m["gap_fte"] = m["effective_capacity"] - m["required_fte"]
     m["understaffed"] = m["gap_fte"] < -1e-9
-    req = m["required_fte"].to_numpy(); capv = m["effective_capacity"].to_numpy()
+    req = m["required_fte"].to_numpy(dtype=float); capv = m["effective_capacity"].to_numpy(dtype=float)
     m["coverage_ratio"] = np.divide(capv, req, out=np.full_like(capv, np.nan), where=req > 1e-9)
-    avail_prod = capv * BUCKET_HOURS * (1.0 - m["shrinkage"].to_numpy())
-    wl = m["workload_hours"].to_numpy()
+    avail_prod = capv * BUCKET_HOURS * (1.0 - m["shrinkage"].to_numpy(dtype=float))
+    wl = m["workload_hours"].to_numpy(dtype=float)
     occ = np.divide(wl, avail_prod, out=np.zeros_like(wl), where=avail_prod > 1e-9)
     occ = np.where((avail_prod <= 1e-9) & (wl > 1e-9), np.inf, occ)
     m["real_occupancy"] = occ
